requirements.txt
langchain~=0.2.1
loguru~=0.7.2
gradio~=4.32.1
python-dotenv~=1.0.1
unstructured[all-docs]~=0.14.3
langchain-chroma~=0.1.1
langchain-community~=0.2.1
langchain-openai~=0.1.7